requirements.txt
# altair
# numpy
# pandas
# pydeck

# ****************requirements.txt****************

streamlit
openai
langchain
faiss-cpu
tiktoken
moviepy
streamlit-chat
ffmpeg_python==0.2.0
librosa==0.9.2
mutagen==1.47.0
pandas==2.0.3
Pillow==10.0.1
Pillow==10.1.0
pydub==0.25.1
python-dotenv==1.0.0
pytube==15.0.0
SpeechRecognition==3.10.0
tqdm==4.66.1
youtube_search_python==1.6.6
youtube_transcript_api==0.6.1
yt_dlp==2023.10.13